adicionando log

+++ python_backend/main.py
@@ -173,8 +173,8 @@ def detect_lines_and_save(image, image_name, h, w, config, current_client):
     for rect in non_overlapping_rectangles:
         x1, y1, x2, y2 = rect
         
-        y1 = int(y1) - int(float(h*0.05))
-        y2 = int(y2) + int(float(h*0.05))
+        y1 = int(y1) - int(h*0.05)
+        y2 = int(y2) + int(h*0.05)
         x1 = int(x1) - int(5)
         x2 = int(x2) + int(5)
